Delay call to file-relative-name until needed
ENCODE_FILE and DECODE_FILE, present in many file and filename
manipulation primitives, are quite expensive and allocate a lot, even
when names are pure ASCII.

We don't actually need to call file-relative-name until there is a
diagnostic emitted for that file, so evaluate it lazily.

This should really be fixed in Emacs, but meanwhile this mitigation
doesn't hurt.