Add checking of ineffective backslashes in string literals
This is an experimental feature.

Check all string literals (not just those that are regexps) for
backslashes that do not escape something sensible; this is a frequent
source of mistakes for regexps in particular, and can indicate errors
in other strings as well.

Currently, all such occurrencies in doc strings are ignored since
those are mostly false positives.